Move fetchWithTimeout to separate module and update usage in subscriptions
